# improve performance of cabal revision merging slightly