*Opened PR on behalf of @cleverca22 to discuss performance improvements*.