@ElvishJerricco 2019-01-27
@cleverca22, @angerman, could that explanation be added as a comment? And the existing comment about preferring r0 should probably be moved accordingly.