@angerman 2019-01-30
before:
```
nix-build release.nix -A  --show-trace  40.94s user 6.87s system 84% cpu 56.698 total
nix-build release.nix -A  --show-trace  40.86s user 6.76s system 68% cpu 1:09.78 total
nix-build release.nix -A  --show-trace  41.97s user 7.08s system 79% cpu 1:02.04 total
nix-build release.nix -A  --show-trace  41.81s user 7.24s system 77% cpu 1:03.54 total
nix-build release.nix -A  --show-trace  41.03s user 7.16s system 81% cpu 59.472 total
```
after
```
nix-build release.nix -I -A  --show-trace  41.62s user 7.18s system 67% cpu 1:12.04 total
nix-build release.nix -I -A  --show-trace  42.12s user 7.23s system 66% cpu 1:14.39 total
nix-build release.nix -I -A  --show-trace  41.28s user 7.20s system 83% cpu 57.958 total
nix-build release.nix -I -A  --show-trace  40.75s user 7.01s system 57% cpu 1:23.03 total
nix-build release.nix -I -A  --show-trace  41.58s user 7.31s system 84% cpu 58.121 total
```

timings are all over the place. But at least it didn't break my build.