@angerman 2019-01-30
@cleverca22 can we get some form of benchmarking/testing for this as well?