@angerman 2019-10-21
I'll close this PR for now. master has evolved too far beyond this.